LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/input-valueasdate-expected.txt

input_valueAsDate


PASS valueAsDate getter on type date (actual value: , expected valueAsDate: null) 
PASS valueAsDate getter on type date (actual value: 0000-12-10, expected valueAsDate: null) 
PASS valueAsDate getter on type date (actual value: 2019-00-12, expected valueAsDate: null) 
PASS valueAsDate getter on type date (actual value: 2019-12-00, expected valueAsDate: null) 
PASS valueAsDate getter on type date (actual value: 2019-13-10, expected valueAsDate: null) 
PASS valueAsDate getter on type date (actual value: 2019-02-29, expected valueAsDate: null) 
FAIL valueAsDate getter on type date (actual value: 2019-12-10, expected valueAsDate: Mon Dec 09 2019 16:00:00 GMT-0800 (PST)) assert_equals: expected object "Mon Dec 09 2019 16:00:00 GMT-0800 (PST)" but got null
FAIL valueAsDate getter on type date (actual value: 2016-02-29, expected valueAsDate: Sun Feb 28 2016 16:00:00 GMT-0800 (PST)) assert_equals: expected object "Sun Feb 28 2016 16:00:00 GMT-0800 (PST)" but got null
FAIL valueAsDate setter on type date (actual valueAsDate: Mon Dec 09 2019 16:00:00 GMT-0800 (PST), expected value: 2019-12-10) The object is in an invalid state.
FAIL valueAsDate setter on type date (actual valueAsDate: Sun Feb 28 2016 16:00:00 GMT-0800 (PST), expected value: 2016-02-29) The object is in an invalid state.
PASS valueAsDate getter on type month (actual value: , expected valueAsDate: null) 
PASS valueAsDate getter on type month (actual value: 0000-12, expected valueAsDate: null) 
PASS valueAsDate getter on type month (actual value: 2019-00, expected valueAsDate: null) 
FAIL valueAsDate getter on type month (actual value: 2019-12, expected valueAsDate: Sat Nov 30 2019 16:00:00 GMT-0800 (PST)) assert_equals: expected object "Sat Nov 30 2019 16:00:00 GMT-0800 (PST)" but got null
FAIL valueAsDate setter on type month (actual valueAsDate: Sat Nov 30 2019 16:00:00 GMT-0800 (PST), expected value: 2019-12) The object is in an invalid state.
PASS valueAsDate getter on type week (actual value: , expected valueAsDate: null) 
PASS valueAsDate getter on type week (actual value: 0000-W50, expected valueAsDate: null) 
PASS valueAsDate getter on type week (actual value: 2019-W00, expected valueAsDate: null) 
PASS valueAsDate getter on type week (actual value: 2019-W60, expected valueAsDate: null) 
FAIL valueAsDate getter on type week (actual value: 2019-W50, expected valueAsDate: Sun Dec 08 2019 16:00:00 GMT-0800 (PST)) assert_equals: expected object "Sun Dec 08 2019 16:00:00 GMT-0800 (PST)" but got null
FAIL valueAsDate setter on type week (actual valueAsDate: Sun Dec 08 2019 16:00:00 GMT-0800 (PST), expected value: 2019-W50) The object is in an invalid state.
PASS valueAsDate getter on type time (actual value: , expected valueAsDate: null) 
PASS valueAsDate getter on type time (actual value: 24:00, expected valueAsDate: null) 
PASS valueAsDate getter on type time (actual value: 00:60, expected valueAsDate: null) 
FAIL valueAsDate getter on type time (actual value: 00:00, expected valueAsDate: Wed Dec 31 1969 16:00:00 GMT-0800 (PST)) assert_equals: expected object "Wed Dec 31 1969 16:00:00 GMT-0800 (PST)" but got null
FAIL valueAsDate getter on type time (actual value: 12:00, expected valueAsDate: Thu Jan 01 1970 04:00:00 GMT-0800 (PST)) assert_equals: expected object "Thu Jan 01 1970 04:00:00 GMT-0800 (PST)" but got null
FAIL valueAsDate getter on type time (actual value: 23:59, expected valueAsDate: Thu Jan 01 1970 15:59:00 GMT-0800 (PST)) assert_equals: expected object "Thu Jan 01 1970 15:59:00 GMT-0800 (PST)" but got null
FAIL valueAsDate setter on type time (actual valueAsDate: Wed Dec 31 1969 16:00:00 GMT-0800 (PST), expected value: 00:00) The object is in an invalid state.
FAIL valueAsDate setter on type time (actual valueAsDate: Thu Jan 01 1970 04:00:00 GMT-0800 (PST), expected value: 12:00) The object is in an invalid state.
FAIL valueAsDate setter on type time (actual valueAsDate: Thu Jan 01 1970 15:59:00 GMT-0800 (PST), expected value: 23:59) The object is in an invalid state.
      
